src/pyproject.toml
[tool.poetry]
name = "fastapi_test_task"
version = "0.1.0"
description = ""
authors = ["mitrandir-pl <leshkopavel@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.89.1"
uvicorn = "^0.20.0"
sqlmodel = "^0.0.8"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.5"
pydantic = {extras = ["email"], version = "^1.10.4"}
PyJWT = "^2.6.0"
passlib = "^1.7.4"
pytest = "^7.2.1"
httpx = "^0.23.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
